requirements.txt
aiogram==2.24
Pillow==9.3.0
requests==2.28.2
regex==2022.10.31
beautifulsoup4==4.11.1
psycopg2==2.9.5